+++ samples/Sentry.Samples.OpenTelemetry.Console/Program.cs
@@ -15,31 +15,37 @@ var activitySource = new ActivitySource("Sentry.Samples.OpenTelemetry.Console");
 SentrySdk.Init(options =>
 {
     // options.Dsn = "... Your DSN ...";
+    options.Dsn = "https://eb18e953812b41c3aeb042e666fd3b5c@o447951.ingest.us.sentry.io/5428537";
     options.TracesSampleRate = 1.0;
     options.UseOpenTelemetry(); // <-- Configure Sentry to use OpenTelemetry trace information
+    options.Debug = true;
 });
 
 using var tracerProvider = Sdk.CreateTracerProviderBuilder()
     .AddSource(activitySource.Name)
+    .AddHttpClientInstrumentation()
     .AddSentry() // <-- Configure OpenTelemetry to send traces to Sentry
     .Build();
 
-// Finally we can use OpenTelemetry to instrument our code. These activities will be captured as a Sentry transaction.
-using var activity = activitySource.StartActivity("Main");
 Console.WriteLine("Hello World!");
-using (var task = activitySource.StartActivity("Task 1"))
+
+// Finally we can use OpenTelemetry to instrument our code. This activity will be captured as a Sentry transaction.
+using (var activity = activitySource.StartActivity("Main"))
 {
-    task?.SetTag("Answer", 42);
-    Thread.Sleep(100); // simulate some work
-    Console.WriteLine("Task 1 completed");
-    task?.SetStatus(Status.Ok);
+    // This creates a span called "Task 1" within the transaction
+    using (var task = activitySource.StartActivity("Task 1"))
+    {
+        task?.SetTag("Answer", 42);
+        Thread.Sleep(100); // simulate some work
+        Console.WriteLine("Task 1 completed");
+        task?.SetStatus(Status.Ok);
+    }
+
+    // Since we use `AddHttpClientInstrumentation` when initializing OpenTelemetry, the following Http request will also
+    // be captured as a Sentry span
+    var httpClient = new HttpClient();
+    var html = await httpClient.GetStringAsync("https://example.com/");
+    Console.WriteLine(html);
 }
 
-using (var task = activitySource.StartActivity("Task 2"))
-{
-    task?.SetTag("Question", "???");
-    Thread.Sleep(100); // simulate some more work
-    Console.WriteLine("Task 2 unresolved");
-    task?.SetStatus(Status.Error);
-}
 Console.WriteLine("Goodbye cruel world...");